Refactored Notifications, added service and repo for notification
